Split legacy into multiple files
